Rename lib/shared/info.js -> lib/shared/runtime-info.js
