+++ lib/cli.js
@@ -21,7 +21,7 @@ const fs = require("fs"),
     { CLIEngine } = require("./cli-engine"),
     options = require("./options"),
     log = require("./shared/logging"),
-    info = require("./shared/info");
+    RuntimeInfo = require("./shared/runtime-info");
 
 const debug = require("debug")("eslint:cli");
 
@@ -163,10 +163,10 @@ const cli = {
         const useStdin = typeof text === "string";
 
         if (currentOptions.version) {
-            log.info(info.version());
+            log.info(RuntimeInfo.version());
         } else if (currentOptions.info) {
             try {
-                log.info(info.environment());
+                log.info(RuntimeInfo.environment());
                 return 0;
             } catch (err) {
                 log.error(err.message);
