+++ lib/shared/runtime-info.js
